Rename bookQuery.ts
