+++ stats/bookEventsByBookQuery.ts
